revert docker + call auth api on product create update and delete

+++ product-api/src/app.controller.ts
@@ -13,11 +13,17 @@ import {
   ProductCRUDServiceController,
   ProductCRUDServiceControllerMethods,
 } from './stubs/product/v1alpha/product';
+import { Metadata, status as RpcStatus } from '@grpc/grpc-js';
+import { AuthService } from './auth/auth.service';
+import { RpcException } from '@nestjs/microservices';
 
 @Controller()
 @ProductCRUDServiceControllerMethods()
 export class AppController implements ProductCRUDServiceController {
-  constructor(private readonly appService: AppService) {}
+  constructor(
+    private readonly appService: AppService,
+    private readonly authService: AuthService,
+  ) {}
   async get(request: GetRequest): Promise<GetResponse> {
     let products: Product[];
 
@@ -31,8 +37,16 @@ export class AppController implements ProductCRUDServiceController {
     return { products: products?.filter((n) => n !== null) ?? [] };
   }
 
-  async update(request: UpdateRequest): Promise<UpdateResponse> {
+  async update(request: UpdateRequest, md: Metadata): Promise<UpdateResponse> {
     const { id, name, description, price, quantity } = request;
+    const user = await this.authService.validate(md);
+    if (!user) {
+      throw new RpcException({
+        code: RpcStatus.UNAUTHENTICATED,
+        message: 'Invalid token',
+      });
+    }
+
     const product = await this.appService.update(id, {
       name,
       description,
@@ -42,20 +56,39 @@ export class AppController implements ProductCRUDServiceController {
     return { product };
   }
 
-  async delete(request: DeleteRequest): Promise<DeleteResponse> {
+  async delete(request: DeleteRequest, md: Metadata): Promise<DeleteResponse> {
     const { id } = request;
+    const user = await this.authService.validate(md);
+    if (!user) {
+      throw new RpcException({
+        code: RpcStatus.UNAUTHENTICATED,
+        message: 'Invalid token',
+      });
+    }
+
     const product = await this.appService.delete(id);
     return { product };
   }
 
-  async add(request: AddRequest): Promise<AddResponse> {
+  async add(request: AddRequest, md: Metadata): Promise<AddResponse> {
     const { name, description, price, quantity } = request;
+
+    const user = await this.authService.validate(md);
+    if (!user) {
+      throw new RpcException({
+        code: RpcStatus.UNAUTHENTICATED,
+        message: 'Invalid token',
+      });
+    }
+
     const product = await this.appService.create({
       name,
       description,
       price,
       quantity,
+      userId: user.userId,
     });
+
     return { product };
   }
 }
